// src/types.ts
export interface WineData {
  Alcohol: number;
  'Malic Acid': number;
  Ash: number;
  'Alcalinity of ash': number;
  Magnesium: number;
  'Total phenols': number;
  Flavanoids: number;
  'Nonflavanoid phenols': number;
  Proanthocyanins:number | string;
  'Color intensity': number;
  Hue: number;
  'OD280/OD315 of diluted wines': number | string;
  Unknown: number;
}






  